+++ CHANGES.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 
 * add a new plugin audiobridge
 * The APIs of Videoroom is compatible with Janus-gateway of v1.1.3
+* support the new cascade mode for videoroom plugin
+* support to subscribe the streams of the different publishers in cascade mode
 
 
  [v1.0.0]  - 2022-07-23